+++ src/models/Model.ts
@@ -1,0 +1,7 @@
+export interface IModel {
+    getTableName(): string;
+}
+
+export default abstract class Model implements IModel {
+    public abstract getTableName(): string;
+}
