+++ src/routes/policy/compile.js
@@ -73,7 +73,7 @@ var RuleCompile = require('../../models/policy/rule_compile');
 var PolicyScript = require('../../models/policy/policy_script');
 
 const config = require('../../config/config');
-const firewallModel = require('../../models/firewall/firewall');
+import { Firewall } from '../../models/firewall/Firewall';
 const socketTools = require('../../utils/socket');
 const policy_rModel = require('../../models/policy/policy_r');
 const fwcError = require('../../utils/error_table');
@@ -202,9 +202,9 @@ router.put('/', (req, res) => {
 			stream.end();
 			
 			// Update firewall status flags.
-			await firewallModel.updateFirewallStatus(req.body.fwcloud,req.body.firewall,"&~1");
+			await Firewall.updateFirewallStatus(req.body.fwcloud,req.body.firewall,"&~1");
 			// Update firewall compile date.
-			await firewallModel.updateFirewallCompileDate(req.body.fwcloud,req.body.firewall);
+			await Firewall.updateFirewallCompileDate(req.body.fwcloud,req.body.firewall);
 
 			socketTools.msgEnd();
 			res.status(204).end();
