+++ src/routes/policy/interface.js
@@ -27,7 +27,7 @@ const policy_r__interfaceModel = require('../../models/policy/policy_r__interfac
 const policy_r__ipobjModel = require('../../models/policy/policy_r__ipobj');
 const policy_rModel = require('../../models/policy/policy_r');
 const policy_cModel = require('../../models/policy/policy_c');
-const firewallModel = require('../../models/firewall/firewall');
+import { Firewall } from '../../models/firewall/Firewall';
 const fwcError = require('../../utils/error_table');
 
 var utilsModel = require("../../utils/utils.js");
@@ -79,7 +79,7 @@ async(req, res) => {
 		// Invalidate compilation of the affected rules.
 		await policy_cModel.deletePolicy_c(rule);
 		await policy_cModel.deletePolicy_c(new_rule);
-		await firewallModel.updateFirewallStatus(req.body.fwcloud,firewall,"|3");
+		await Firewall.updateFirewallStatus(req.body.fwcloud,firewall,"|3");
 
 		// Get positions content.
 		const data = await 	policy_r__ipobjModel.getPositionsContent(req.dbCon, position, new_position);
