+++ src/routes/policy/prefix.js
@@ -29,7 +29,7 @@ const policyPrefixModel = require('../../models/policy/prefix');
 const policy_r__ipobjModel = require('../../models/policy/policy_r__ipobj');
 const policy_rModel = require('../../models/policy/policy_r');
 const policy_cModel = require('../../models/policy/policy_c');
-const firewallModel = require('../../models/firewall/firewall');
+import { Firewall } from '../../models/firewall/Firewall';
 const utilsModel = require("../../utils/utils.js");
 const fwcError = require('../../utils/error_table');
 
@@ -71,7 +71,7 @@ async (req, res) => {
 		// Invalidate compilation of the affected rules.
 		await policy_cModel.deletePolicy_c(req.body.rule);
 		await policy_cModel.deletePolicy_c(req.body.new_rule);
-		await firewallModel.updateFirewallStatus(req.body.fwcloud,req.body.firewall,"|3");
+		await Firewall.updateFirewallStatus(req.body.fwcloud,req.body.firewall,"|3");
 
 		// Move OpenVPN configuration object to the new position.
 		const data = await policyPrefixModel.moveToNewPosition(req);
